// todo-webapp/src/generated/todo-api.ts
export interface components {
    schemas: {
        Todo: {
            id: string;
            title: string;
            completed: boolean;
        };
        NewTodo: {
            title: string;
        };
        UpdateTodo: {
            title?: string;
            completed?: boolean;
        };
        Error: {
            /** @description HTTP or application error code */
            code: number;
            /** @description short human-readable label */
            message: string;
            /** @description detailed explanation */
            description?: string;
            /** @description URI to documentation */
            moreInfo?: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}